--- rcjaRegistration/apiv1/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class Apiv1Config(AppConfig):
-    name = 'apiv1'
